Step 8 Complete!
Decided the theme will be grains of Sand and the player being a being who controls things to gather sand.
Added comments to show some of the changes made to connect them to the theme.
Such as changing the names of the buttons and upgrades to fit.
Added 2 new images one for the title and then changed the clicker button to be an image of sand.