Step 9 Complete!
Using the idea to swap over to an array with loops I made my own changes
Overhauled most of the system to use the new system which should now make adding new upgrades easier.
Added some new comments above the changes to explain the general idea of what is happening.